backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anaconda-clean @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_355f5d45-4286-44ed-8ffe-4cfe2dab569cbs6o5bqv/croots/recipe/anaconda-clean_1652723726562/work
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.1.1
argilla==0.0.1
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitarray==2.7.3
black==23.3.0
bleach==6.0.0
blessed==1.20.0
boltons @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f63n9uulmp/croot/boltons_1677628710094/work
brotlipy==0.7.0
bs4==0.0.1
build==1.0.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ab19r4bji3/croot/cffi_1670423206034/work
charset-normalizer==3.1.0
chroma-hnswlib==0.7.2
chromadb==0.4.8
click==8.1.3
click-log==0.4.0
clickhouse-connect==0.5.20
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
conda @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3bspr1yk3p/croot/conda_1685025187626/work
conda-content-trust @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11146a2b-93c2-444c-a378-ad4fac363e991s0r1hnp/croots/recipe/conda-content-trust_1658126383571/work
conda-package-handling @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5f3f459f9v/croot/conda-package-handling_1672865025324/work
conda_package_streaming @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_58gpsa_6af/croot/conda-package-streaming_1670508144037/work
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cytoolz==0.12.1
dataclasses-json==0.5.7
Deprecated==1.2.13
deprecation==2.1.0
diskcache==5.6.1
dnspython==2.3.0
docutils==0.20.1
dotty-dict==1.3.1
duckdb==0.7.1
duckduckgo-search==2.9.3
effdet==0.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.1.3
faiss-cpu==1.7.3
fastapi==0.99.1
filelock==3.12.0
firebase-admin==6.1.0
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
fvcore==0.1.5.post20221221
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
gotrue==1.0.4
gptcache==0.1.11
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
hexbytes==0.3.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inquirer==3.1.3
invoke==1.7.3
iopath==0.1.9
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.17.3
keyring==24.2.0
kiwisolver==1.4.4
kor==0.9.2
kor-project==0.2.1
langchain==0.0.202
langchainplus-sdk==0.0.20
langsmith==0.0.33
layoutparser==0.3.4
llama-index==0.6.9
lru-dict==1.1.8
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
open-interpreter==0.1.1
openai==0.27.10
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
orjson==3.8.11
overrides==7.4.0
packaging==23.1
pandas==1.5.3
parsimonious==0.9.0
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pip-tools==7.3.0
pipx==1.2.0
pkginfo==1.9.6
platformdirs==3.2.0
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-w6jyveby/pluggy_1648109277227/work
portalocker==2.7.0
postgrest==0.10.8
postgrest-py==0.10.3
posthog==3.0.0
proto-plus==1.22.2
protobuf==4.22.3
pulsar-client==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e9xhel3kyj/croot/pycosat_1666805516297/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.17
pydantic==1.10.9
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.15.0
PyJWT==2.7.0
pyOpenSSL @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_38h4axtq38/croot/pyopenssl_1677607699670/work
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks @ file:///Users/ktietz/ci_310/pysocks_1643961536721/work
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.0
python-editor==1.0.4
python-gitlab==3.15.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-semantic-release==7.33.0
python-telegram-bot==20.3
pytube==12.1.3
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
readme-renderer==41.0
realtime==1.0.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.5.2
rlp==3.0.0
rsa==4.9
ruamel.yaml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_aeh5mqcw49/croot/ruamel.yaml_1666304555976/work
ruamel.yaml.clib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f64xdg2rww/croot/ruamel.yaml.clib_1666302244208/work
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
starlette==0.27.0
storage3==0.5.0
StrEnum==0.4.15
supabase==1.0.0
supafunc==0.2.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tiktoken==0.4.0
timm==0.6.13
tokenizers==0.13.3
tokentrim==0.1.10
tomli==2.0.1
tomlkit==0.12.1
toolz @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_362wyqvvgy/croot/toolz_1667464079070/work
torch==2.0.0
torchvision==0.15.1
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ac7zic_tin/croot/tqdm_1679561870178/work
transformers==4.28.1
twilio==8.5.0
twine==3.8.0
typing-inspect==0.8.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.7.0
unstructured==0.5.9
unstructured-inference==0.2.4
uritemplate==4.1.1
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_c7xh0_scrc/croot/urllib3_1680254693282/work
userpath==1.8.0
uvicorn==0.21.1
uvloop==0.17.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
web3==6.2.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
wget==3.2
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.14.1
XlsxWriter==3.1.0
xmltodict==0.13.0
yacs==0.1.8
yarl==1.8.2
youtube-transcript-api==0.5.0
zipp==3.16.2
zstandard==0.21.0
